pyproject.toml
[tool.poetry]
name = "walrus"
version = "0.1.0"
description = "An onchain AI agent"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
coinbase-agentkit = "^0.4.0"
langchain-openai = "^0.2.4"
langgraph = "^0.2.39"
coinbase-agentkit-langchain = "0.3.0"

[tool.poetry.scripts]
start-agent = "chatbot:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
